fat: add implementation of modules/esm
